Overhaul data pipeline and some other stuff idk
- I overhauled the data pipeline to scrape all the player/position data
from PFR instead of building it from the nflverse play-by-play data.
This is because I was having a difficult time replicating the official
stats, and some things in the pbp data were making little sense to me.
- I added permutation importances for feature evaluation.
- The model now knows which QB played in each game and what their
cumulative passer rating is over that QB's last n games.
- Some other stuff. Mostly minor cleanup. Nobody will ever read this
commit message ever again, so it's fine.